feat: Bairro services and controller + clean some codes

+++ ecommerce/src/main/java/com/gustavo/ecommerce/repository/BairroRepository.java
@@ -2,6 +2,7 @@ package com.gustavo.ecommerce.repository;
 
 import com.gustavo.ecommerce.entity.Bairro;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
 
-public interface BairroRepository extends JpaRepository<Bairro, Long> {
+public interface BairroRepository extends JpaRepository<Bairro, Integer> {
 }
